Installing Radamsa
Added the capability to install Radamsa. Radamsa is a test case generator for robustness testing. It is typically used to test how well a program can withstand malformed and potentially malicious inputs. It works by reading sample files of valid data and generating interestingly different outputs from them. The main selling points of `radamsa` are that it has already found a slew of bugs in programs that actually matter, it is easily scriptable and, easy to get up and running.